# docs/api-spec.yaml
openapi: 3.0.3
info:
  title: AI Research War-Room API
  description: API for orchestrating multi-agent AI debates for research and startup idea refinement.
  version: 1.0.0
servers:
  - url: http://localhost:8080
    description: Local development server
paths:
  /api/projects:
    get:
      summary: List all projects
      responses:
        '200':
          description: Successful retrieval
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/ProjectResponse'
    post:
      summary: Create a new project
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ProjectRequest'
      responses:
        '201':
          description: Project created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProjectResponse'
        '400':
          description: Invalid input
  /api/projects/{id}:
    parameters:
      - name: id
        in: path
        required: true
        schema:
          type: string
          format: uuid
    get:
      summary: Get project by ID
      responses:
        '200':
          description: Project details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProjectResponse'
        '404':
          description: Project not found
    put:
      summary: Update project
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ProjectRequest'
      responses:
        '200':
          description: Project updated
    delete:
      summary: Delete project
      responses:
        '204':
          description: Project deleted
  /api/upload:
    post:
      summary: Upload research paper (PDF)
      requestBody:
        content:
          multipart/form-data:
            schema:
              type: object
              properties:
                file:
                  type: string
                  format: binary
      responses:
        '200':
          description: Cleaned text extracted from PDF
          content:
            text/plain:
              schema:
                type: string
components:
  schemas:
    ProjectRequest:
      type: object
      required:
        - title
        - description
        - type
      properties:
        title:
          type: string
        description:
          type: string
        type:
          type: string
          enum: [RESEARCH, STARTUP, ML]
    ProjectResponse:
      type: object
      properties:
        id:
          type: string
          format: uuid
        title:
          type: string
        description:
          type: string
        type:
          type: string
        status:
          type: string
        finalContent:
          type: string
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
    ErrorResponse:
      type: object
      properties:
        message:
          type: string
        code:
          type: string
        timestamp:
          type: string
          format: date-time